==> 01 Clean.ipynb <==
### Data Cleaning

#### Import Zone

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

#### View df_final_web_data_pt_1

# df_final_web_data_pt_1
path_1 = "C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_web_data_pt_1.txt"
df_final_web_data_pt_1 = pd.read_csv(path_1)
df_final_web_data_pt_1.info()

df_final_web_data_pt_1.nunique()

# Explore the size of the dataset
df_final_web_data_pt_1.shape

# Checking for Null Values
df_final_web_data_pt_1.isnull().value_counts()

# Check for null values in each column
df_final_web_data_pt_1.isna().any()

#### View df_final_web_data_pt_2

# df_final_web_data_pt_2
path_2 = "C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_web_data_pt_2.txt"
df_final_web_data_pt_2 = pd.read_csv(path_2)
df_final_web_data_pt_2.info()

df_final_web_data_pt_2.nunique()

# Explore the size of the dataset
df_final_web_data_pt_2.shape

# Checking for Null Values
df_final_web_data_pt_2.isnull().value_counts()

# Check for null values in each column
df_final_web_data_pt_2.isna().any()

#### View df_final_demo

# df_final_demo
path_3 = "C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_demo.txt"
df_final_demo = pd.read_csv(path_3)
df_final_demo.info()

# Explore the size of the dataset
df_final_demo.shape

# Check for null values in each column
df_final_demo.isna().any()

#number of unique values per column
df_final_demo.nunique()

#### View df_final_experiment_clients

# df_final_experiment_clients
path_4 = "C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_experiment_clients.txt"
df_final_experiment_clients = pd.read_csv(path_4)
df_final_experiment_clients.info()

# Explore the size of the dataset
df_final_experiment_clients.shape

# Check for null values in each column
df_final_experiment_clients.isna().any()

# Number of unique values per column
df_final_experiment_clients.nunique()

#### Concatenating df_final_web_data_pt_1 y df_final_web_data_pt_2

df_concat_final_web_data= pd.concat([df_final_web_data_pt_1,df_final_web_data_pt_2], axis=0)

df_concat_final_web_data['date_time'] = pd.to_datetime(df_concat_final_web_data['date_time'])
df_concat_final_web_data.info()

# 125,821
df_concat_final_web_data.nunique()

# Check for null values in each column
df_concat_final_web_data.isna().any()

#number of unique values per column
df_concat_final_web_data.nunique()

mapping = {
    'start' : '0_start',
    'step_1' : '1_step_1', 
    'step_2' : '2_step_2',
    'step_3' : '3_step_3',
    'confirm' : '4_confirm',
}
# Reemplazamos los datos de M x 2 y los diversos datos de M por "Male" y F por "Female"
df_concat_final_web_data['process_step'] = df_concat_final_web_data['process_step'].replace(mapping)

#### Clean up:

##### df_final_demo

# Deleting: 7402828,355337,8412164,2222915,4876926,5277910,7616759,8191345,1227228,8611797,5144725,1037867,1618749,6021001
df_final_demo[['clnt_tenure_mnth', 'clnt_age', 'gendr', 'num_accts', 'bal', 'calls_6_mnth', 'logons_6_mnth']].isna().any()


df_final_demo.dropna(inplace=True)

df_final_demo.isna().any()

df_final_demo.info()

df_final_demo.shape

df_final_demo['num_accts'] = df_final_demo['num_accts'].astype('Int64')
df_final_demo['calls_6_mnth'] = df_final_demo['calls_6_mnth'].astype('Int64')
df_final_demo['logons_6_mnth'] = df_final_demo['logons_6_mnth'].astype('Int64')


df_final_demo['bal'] = df_final_demo['bal'].round(2)

# Convert gendr with X value for U
df_final_demo['gendr'].replace('X','U',inplace=True)

df_final_demo['gendr'].value_counts()

df_final_demo.info()

##### df_final_experiment_clients

df_final_experiment_clients.dropna(inplace=True)

df_final_experiment_clients.shape

df_final_experiment_clients.isna().any()

#### Export Clean CSV files

##### Clean CSV for concat file df_concat_final_web_data that comes from: df_final_web_data_pt_1, df_final_web_data_pt_2

df_concat_final_web_data.to_csv('C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_concat_final_web_data.csv',index=False,encoding='latin-1',sep=';')

##### Clean csv file for: df_final_experiment_clients

df_final_experiment_clients.to_csv('C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_experiment_clients.csv',index=False,encoding='latin-1',sep=';')

##### Clean csv file for: df_final_demo

df_final_demo.to_csv('C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_demo.csv',index=False,encoding='latin-1',sep=';')
