==> 03 Client_Behaviour.ipynb <==
### Client behavior analysis

#### Import Zone

import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
from functools import reduce

df_final_experiment_clients = pd.read_csv('C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_experiment_clients.csv',sep=';')

df_concat_final_web_data = pd.read_csv('C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_concat_final_web_data.csv',sep=';')

df_final_demo = pd.read_csv('C:/Users/user/Jupyter/IronHack/Work_Python/Week5/MiniProject/df_final_demo.csv',sep=';')

df_final_demo.drop(columns=['Unnamed: 0'], inplace=True)

df_final_experiment_clients.head()

df_final_experiment_clients[df_final_experiment_clients['Variation']=='Control']['client_id'].count(), df_final_experiment_clients[df_final_experiment_clients['Variation']=='Test']['client_id'].count()

#### Who are the primary clients using this online process?


sorted_df = df_final_demo.sort_values(by=['bal'], ascending=False).head(10)
sorted_df

#### Are the primary clients younger or older?

mean_age_all = round(df_final_demo['clnt_age'].mean(),0)
mean_age_primary = round(sorted_df['clnt_age'].mean(),0)
print(f'Older. The average age for the primary clients is: {mean_age_primary} '
      f'\n      higher than the avg age of all clients which is: {mean_age_all}')

#### New or long-standing?

mean_tenure_all = round(df_final_demo['clnt_tenure_mnth'].mean(),0)
mean_tenure_primary = round(sorted_df['clnt_tenure_mnth'].mean(),0)
print(f'Yes. The average tenure for the primary clients is: {mean_tenure_primary} months '
      f'\n      higher than the avg tenure of all clients which is: {mean_tenure_all} months')

### Next, carry out a client behaviour analysis to answer any additional relevant questions you think are important.

print(f'The mean age of the clients is {round((df_final_demo['clnt_age'].mean()),0)} years old')

print(f'The mean calls from clients in the last 6 months is {round((df_final_demo['calls_6_mnth'].mean()),0)} calls')
